fix: stabilize rendering after sculpting and reduce memory footprint
- Refresh bounding sphere and box on capped geometry after live preview
  writes and initial cap build so stale frustum culling cannot make a
  sculpted vessel vanish mid-drag
- Skip non-finite vertex results in the sculpt preview so a degenerate
  plane intersection cannot poison positions, normals, and bounds with
  NaN cascades
- Reset the body cursor on unmount; scope keyboard effect deps to stable
  undo/redo callbacks instead of the per-render history object
- Halve chamfer distance storage to Float32 and drop directional model
  cache limit to eight, trimming worst-case retained memory per vessel

+++ src/lib/metrics/sdf.ts
@@ -209,7 +209,7 @@ export function buildVoxelSdf(
   }
 
   const total = nx * ny * nz;
-  const distanceUnits = new Float64Array(total);
+  const distanceUnits = new Float32Array(total);
   const largeDistance = 1e15;
 
   const statusAt = (i: number, j: number, k: number) =>
